ajout ResponseEntity

+++ src/main/java/edu/fbansept/devlog2021/controller/CompetenceController.java
@@ -1,14 +1,14 @@
 package edu.fbansept.devlog2021.controller;
 
-import com.fasterxml.jackson.annotation.JsonView;
 import edu.fbansept.devlog2021.dao.CompetenceDao;
 import edu.fbansept.devlog2021.model.Competence;
-import edu.fbansept.devlog2021.model.Utilisateur;
-import edu.fbansept.devlog2021.view.CustomJsonView;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @CrossOrigin
@@ -21,38 +21,43 @@ public class CompetenceController {
         this.competenceDao = competenceDao;
     }
 
+    @GetMapping("/user/competence/{id}")
+    public ResponseEntity<Competence> getCompetence(@PathVariable int id) {
 
-    @GetMapping("/admin/competence/{id}")
-    @JsonView(CustomJsonView.VueCompetence.class)
-    public Competence getCompetence(@PathVariable int id) {
-        return competenceDao.findById(id).orElse(null);
+        Optional<Competence> competence = competenceDao.findById(id);
+
+        if(competence.isPresent()) {
+            return ResponseEntity.ok(competence.get());
+        } else {
+            return ResponseEntity.noContent().build();
+        }
     }
 
-    @GetMapping("/admin/competences")
-    @JsonView(CustomJsonView.VueCompetence.class)
-    public List<Competence> getCompetences () {
+    @GetMapping("/user/competences")
+    public ResponseEntity<List<Competence>> getCompetences () {
 
-        return competenceDao.findAll();
-
+        return ResponseEntity.ok(competenceDao.findAll());
     }
 
     @PostMapping("/admin/competence")
-    public boolean addCompetence (@RequestBody Competence competence) {
+    public ResponseEntity<String> addCompetence (@RequestBody Competence competence) {
 
-        competenceDao.save(competence);
-
-        return true;
+        competence = competenceDao.saveAndFlush(competence);
+        return ResponseEntity.created(
+                URI.create("/user/competence/" + competence.getId())
+        ).build();
     }
 
     @DeleteMapping("/admin/competence/{id}")
-    public boolean deleteCompetence (@PathVariable int id) {
+    public ResponseEntity<Integer> deleteCompetence (@PathVariable int id) {
 
-        competenceDao.deleteById(id);
-
-        return true;
+        if(competenceDao.existsById(id)) {
+            competenceDao.deleteById(id);
+            return ResponseEntity.ok(id);
+        } else {
+            return ResponseEntity.noContent().build();
+        }
     }
-
-
 }
