+++ src/main/java/edu/fbansept/devlog2021/controller/StatutController.java
@@ -3,9 +3,12 @@ package edu.fbansept.devlog2021.controller;
 import edu.fbansept.devlog2021.dao.StatutDao;
 import edu.fbansept.devlog2021.model.Statut;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @CrossOrigin
@@ -18,36 +21,43 @@ public class StatutController {
         this.statutDao = statutDao;
     }
 
-    @GetMapping("/admin/statut/{id}")
-    public Statut getStatut(@PathVariable int id) {
-        return statutDao.findById(id).orElse(null);
+    @GetMapping("/user/statut/{id}")
+    public ResponseEntity<Statut> getStatut(@PathVariable int id) {
+
+        Optional<Statut> statut = statutDao.findById(id);
+
+        if(statut.isPresent()) {
+            return ResponseEntity.ok(statut.get());
+        } else {
+            return ResponseEntity.noContent().build();
+        }
     }
 
-    @GetMapping("/admin/statuts")
-    public List<Statut> getStatuts () {
+    @GetMapping("/user/statuts")
+    public ResponseEntity<List<Statut>> getStatuts () {
 
-        List<Statut> listeStatut = statutDao.findAll();
-
-        return listeStatut;
+        return ResponseEntity.ok(statutDao.findAll());
     }
 
     @PostMapping("/admin/statut")
-    public boolean addStatut (@RequestBody Statut statut) {
+    public ResponseEntity<String> addStatut (@RequestBody Statut statut) {
 
-        statutDao.save(statut);
-
-        return true;
+        statut = statutDao.saveAndFlush(statut);
+        return ResponseEntity.created(
+                URI.create("/user/statut/" + statut.getId())
+        ).build();
     }
 
     @DeleteMapping("/admin/statut/{id}")
-    public boolean deleteStatut (@PathVariable int id) {
+    public ResponseEntity<Integer> deleteStatut (@PathVariable int id) {
 
-        statutDao.deleteById(id);
-
-        return true;
+        if(statutDao.existsById(id)) {
+            statutDao.deleteById(id);
+            return ResponseEntity.ok(id);
+        } else {
+            return ResponseEntity.noContent().build();
+        }
     }
-
-
 }
